Label for first vertical line and some refactoring
