+++ lab-2/src/jvmMain/kotlin/io/github/e1turin/output/view/shared/lib/plot/units.kt
@@ -1,7 +1,17 @@
 package io.github.e1turin.output.view.shared.lib.plot
 
+import androidx.compose.runtime.Stable
 import io.github.e1turin.output.view.entities.plot.model.Gu
-import io.github.e1turin.output.view.entities.plot.ui.gu
 
 fun ClosedRange<Double>.toGu(): ClosedRange<Gu> =
     start.gu..endInclusive.gu
+
+@Stable
+inline val Float.gu: Gu get() = Gu(this)
+
+@Stable
+inline val Double.gu: Gu get() = Gu(this.toFloat())
+
+@Stable
+inline val Int.gu: Gu get() = Gu(this.toFloat())
+
